@@ -14,7 +14,7 @@
 	<description>Misc commands to improve administration on server.</description>
 
 	<properties>
-		<java.version>1.8</java.version>
+		<java.version>17</java.version>
 		<pluginName>SZAdminUtilities</pluginName>
 		<pluginId>szadminutilities</pluginId>
 		<author>szumielxd</author>
@@ -23,7 +23,7 @@
 		<package>me.szumielxd.adminutilities</package>
 		<velocityMainClass>${package}.velocity.AdminUtilitiesVelocity</velocityMainClass>
 		<bungeeMainClass>${package}.bungee.AdminUtilitiesBungee</bungeeMainClass>
-		<revision>2.0.1</revision>
+		<revision>2.1.0</revision>
 		<finalName>/../../target/${pluginName}-${project.version}</finalName>
 	</properties>
 
@@ -82,7 +82,7 @@
 			<dependency>
 				<groupId>net.kyori</groupId>
 				<artifactId>adventure-platform-bungeecord</artifactId>
-				<version>4.0.0-SNAPSHOT</version>
+				<version>4.3.1</version>
 				<type>jar</type>
 				<scope>compile</scope>
 				<optional>true</optional>
@@ -128,7 +128,7 @@
 			<dependency>
 				<groupId>net.kyori</groupId>
 				<artifactId>adventure-platform-api</artifactId>
-				<version>4.0.0-SNAPSHOT</version>
+				<version>4.3.1</version>
 				<scope>compile</scope>
 				<exclusions>
 					<exclusion>
@@ -144,7 +144,7 @@
 			<dependency>
 				<groupId>net.kyori</groupId>
 				<artifactId>adventure-text-serializer-bungeecord</artifactId>
-				<version>4.0.0-SNAPSHOT</version>
+				<version>4.3.1</version>
 				<scope>compile</scope>
 				<exclusions>
 					<exclusion>
@@ -160,13 +160,13 @@
 			<dependency>
 				<groupId>net.kyori</groupId>
 				<artifactId>adventure-text-serializer-legacy</artifactId>
-				<version>4.8.1</version>
+				<version>4.14.0</version>
 				<scope>compile</scope>
 			</dependency>
 			<dependency>
 				<groupId>net.kyori</groupId>
 				<artifactId>adventure-text-serializer-gson</artifactId>
-				<version>4.8.1</version>
+				<version>4.14.0</version>
 				<scope>compile</scope>
 				<exclusions>
 					<exclusion>
@@ -187,7 +187,7 @@
 			<dependency>
 				<groupId>net.md-5</groupId>
 				<artifactId>bungeecord-api</artifactId>
-				<version>1.17-R0.1-SNAPSHOT</version>
+				<version>1.20-R0.1-SNAPSHOT</version>
 				<scope>provided</scope>
 			</dependency>
 			
@@ -195,7 +195,7 @@
 			<dependency>
 				<groupId>com.velocitypowered</groupId>
 				<artifactId>velocity-api</artifactId>
-				<version>3.0.1</version>
+				<version>3.3.0-SNAPSHOT</version>
 				<scope>provided</scope>
 			</dependency>
 		</dependencies>
@@ -204,21 +204,21 @@
 	<!-- Dependencies -->
 	<dependencies>
 		<dependency>
-			<groupId>me.szumielxd</groupId>
-			<artifactId>SZLobbySystem</artifactId>
-			<version>1.1.0</version>
-			<scope>provided</scope>
-		</dependency>
-		<dependency>
 			<groupId>net.luckperms</groupId>
 			<artifactId>api</artifactId>
 			<version>5.2</version>
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
+			<groupId>org.projectlombok</groupId>
+			<artifactId>lombok</artifactId>
+			<version>1.18.20</version>
+			<scope>provided</scope>
+		</dependency>
+		<dependency>
 			<groupId>org.jetbrains</groupId>
 			<artifactId>annotations</artifactId>
-			<version>21.0.1</version>
+			<version>22.0.0</version>
 			<scope>provided</scope>
 		</dependency>
 		<dependency>
@@ -240,10 +240,11 @@
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-compiler-plugin</artifactId>
-					<version>3.8.1</version>
+					<version>3.11.0</version>
 					<configuration>
-						<source>1.8</source>
-						<target>1.8</target>
+						<source>17</source>
+						<target>17</target>
+						<release>17</release>
 						<showDeprecation>true</showDeprecation>
 					</configuration>
 				</plugin>
@@ -309,7 +310,7 @@
 				<plugin>
 					<groupId>org.apache.maven.plugins</groupId>
 					<artifactId>maven-shade-plugin</artifactId>
-					<version>3.2.4</version>
+					<version>3.5.1</version>
 					<executions>
 						<execution>
 							<phase>package</phase>
